requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
loadotenv==1.0.1
MarkupSafe==3.0.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
